Rename and fix typo
